+++ java-io-picocsv/src/test/java/nbbrd/io/picocsv/PicocsvTest.java
@@ -7,8 +7,8 @@ import nbbrd.picocsv.Csv;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.Reader;
 import java.io.StringWriter;
 import java.nio.charset.Charset;
 import java.nio.file.Path;
@@ -49,7 +49,7 @@ public class PicocsvTest {
 
         assertThatNullPointerException().isThrownBy(() -> x.parseCsv((Csv.Reader) null));
 
-        try (BufferedReader chars = RESOURCE_ID.open(UTF_8)) {
+        try (Reader chars = RESOURCE_ID.open(UTF_8)) {
             try (Csv.Reader csv = Csv.Reader.of(x.getFormat(), x.getOptions(), chars, DEFAULT_CHAR_BUFFER_SIZE)) {
                 assertThat(x.parseCsv(csv))
                         .containsExactlyElementsOf(USERS);
@@ -66,7 +66,7 @@ public class PicocsvTest {
 
         assertThatNullPointerException().isThrownBy(() -> x.parseCsv((IOSupplier<Csv.Reader>) null));
 
-        try (BufferedReader chars = RESOURCE_ID.open(UTF_8)) {
+        try (Reader chars = RESOURCE_ID.open(UTF_8)) {
             assertThat(x.parseCsv(() -> Csv.Reader.of(x.getFormat(), x.getOptions(), chars, DEFAULT_CHAR_BUFFER_SIZE)))
                     .containsExactlyElementsOf(USERS);
         }
@@ -81,9 +81,13 @@ public class PicocsvTest {
                 .format(UNIX)
                 .build();
 
-        String expected = RESOURCE_ID.copyByLineToString(UTF_8, Csv.Format.UNIX_SEPARATOR);
+        String expected = normalizeNewlines(RESOURCE_ID.copyToString(UTF_8));
 
         assertTextFormatterCompliance(temp, x, USERS, encoding -> expected, ENCODINGS);
+    }
+
+    private static String normalizeNewlines(CharSequence actual) {
+        return actual.toString().replace("\r\n", "\n");
     }
 
     @Test
@@ -96,13 +100,12 @@ public class PicocsvTest {
         assertThatNullPointerException().isThrownBy(() -> x.formatCsv(null, Csv.Writer.of(Csv.Format.DEFAULT, Csv.WriterOptions.DEFAULT, new StringWriter(), DEFAULT_CHAR_BUFFER_SIZE)));
         assertThatNullPointerException().isThrownBy(() -> x.formatCsv(Collections.emptyList(), (Csv.Writer) null));
 
-        String expected = RESOURCE_ID.copyByLineToString(UTF_8, Csv.Format.UNIX_SEPARATOR);
-
         try (StringWriter chars = new StringWriter()) {
             try (Csv.Writer csv = Csv.Writer.of(x.getFormat(), x.getOptions(), chars, DEFAULT_CHAR_BUFFER_SIZE)) {
                 x.formatCsv(USERS, csv);
             }
-            assertThat(chars.toString()).isEqualTo(expected);
+            assertThat(chars.toString())
+                    .isEqualToNormalizingNewlines(RESOURCE_ID.copyToString(UTF_8));
         }
     }
 
@@ -116,11 +119,10 @@ public class PicocsvTest {
         assertThatNullPointerException().isThrownBy(() -> x.formatCsv(null, Csv.Writer.of(Csv.Format.DEFAULT, Csv.WriterOptions.DEFAULT, new StringWriter(), DEFAULT_CHAR_BUFFER_SIZE)));
         assertThatNullPointerException().isThrownBy(() -> x.formatCsv(Collections.emptyList(), (IOSupplier<Csv.Writer>) null));
 
-        String expected = RESOURCE_ID.copyByLineToString(UTF_8, Csv.Format.UNIX_SEPARATOR);
-
         try (StringWriter chars = new StringWriter()) {
             x.formatCsv(USERS, () -> Csv.Writer.of(x.getFormat(), x.getOptions(), chars, DEFAULT_CHAR_BUFFER_SIZE));
-            assertThat(chars.toString()).isEqualTo(expected);
+            assertThat(chars.toString())
+                    .isEqualToNormalizingNewlines(RESOURCE_ID.copyToString(UTF_8));
         }
     }