+++ java-io-xml-bind/src/main/java/nbbrd/io/xml/bind/Jaxb.java
@@ -154,7 +154,7 @@ public class Jaxb {
         }
 
         private static XMLInputFactory getStaxFactory() {
-            XMLInputFactory result = XMLInputFactory.newFactory();
+            XMLInputFactory result = XMLInputFactory.newInstance();
             Stax.preventXXE(result);
             return result;
         }
@@ -192,7 +192,7 @@ public class Jaxb {
         }
 
         private static Object parseFileXXE(Unmarshaller engine, File source, XMLInputFactory xxe) throws IOException {
-            try (InputStream resource = LegacyFiles.newInputStream(source)) {
+            try (InputStream resource = LegacyFiles.openInputStream(source)) {
                 XMLStreamReader reader = xxe.createXMLStreamReader(LegacyFiles.toSystemId(source), resource);
                 try {
                     return engine.unmarshal(reader);